Update README.md
ass.7
i made the changes for part 1 and 2 in DEMO file

1- try catch , finaly for getInstance 2- try , catch and finaly for send 3- try , catch and finaly for release

in part 2 added the loger to the try cathc statment and system.oyt .print in all project

Explain why the logger uses a singleton design pattern. It is utilized when single one instance of a class is needed to control the operation throughout the execution.
In each scenario and at any cost, a singleton class should not contain numerous instances. For logging, driver objects, caching and thread pooling, and database connections, singleton classes are employed.